pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "youtube-summarizer-cloud-run"
version = "0.1.0"
description = "Tutorial on developing a Flask application to summarize YouTube videos with Cloud Run"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = ["Yury Kashnitsky <kashnitsky@google.com>"]
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.2.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.2",
]
